+++ src/main/scala/edu/ie3/simona/model/thermal/ThermalGrid.scala
@@ -13,6 +13,7 @@ import edu.ie3.datamodel.models.result.thermal.{
   CylindricalStorageResult,
   ThermalHouseResult,
 }
+import edu.ie3.simona.exceptions.InvalidParameterException
 import edu.ie3.simona.exceptions.agent.InconsistentStateException
 import edu.ie3.simona.model.participant.HpModel.{HpRelevantData, HpState}
 import edu.ie3.simona.model.thermal.ThermalGrid.{
@@ -30,6 +31,7 @@ import squants.{Energy, Power, Temperature}
 
 import java.time.ZonedDateTime
 import scala.jdk.CollectionConverters.SetHasAsScala
+import scala.language.postfixOps
 
 /** Calculation model for a thermal grid. It is assumed, that all elements are
   * connected directly with exactly one thermal bus
@@ -147,82 +149,345 @@ final case class ThermalGrid(
     *
     * @param relevantData
     *   data of heat pump including state of the heat pump
-    * @param state
-    *   Currently applicable state
+    * @param lastThermalGridState
+    *   state of the thermalGrid until this tick
     * @param lastAmbientTemperature
     *   Ambient temperature valid up until (not including) the current tick
+    * @param isRunning
+    *   determines whether the heat pump is running or not
     * @param qDot
-    *   Thermal energy balance
+    *   Infeed to the grid from thermal generation (e.g. heat pump) or thermal
+    *   storages
+    * @param thermalDemands
+    *   holds the thermal demands of the thermal units (house, storage)
     * @return
     *   The updated state of the grid
     */
   def updateState(
       relevantData: HpRelevantData,
-      state: ThermalGridState,
+      lastThermalGridState: ThermalGridState,
       lastAmbientTemperature: Temperature,
+      isRunning: Boolean,
       qDot: Power,
+      thermalDemands: ThermalDemandWrapper,
   ): (ThermalGridState, Option[ThermalThreshold]) = if (qDot > zeroKW)
-    handleInfeed(relevantData, lastAmbientTemperature, state, qDot)
+    handleInfeed(
+      relevantData,
+      lastAmbientTemperature,
+      lastThermalGridState,
+      isRunning,
+      qDot,
+      thermalDemands,
+    )
   else
     handleConsumption(
       relevantData,
       lastAmbientTemperature,
-      state,
+      lastThermalGridState,
       qDot,
     )
 
-  /** Handles the case, when a grid has infeed. First, heat up all the houses to
-    * their maximum temperature, then fill up the storages
+  /** Handles the case, when a grid has infeed. Depending on which entity has
+    * some heat demand the house or the storage will be heated up / filled up.
+    * First the actions from lastState will be considered and checked if the
+    * behaviour should be continued. This might be the case, if we got activated
+    * by updated weather data. If this is not the case, all other cases will be
+    * handled by [[ThermalGrid.handleFinalInfeedCases]]
     *
     * @param relevantData
     *   data of heat pump including state of the heat pump
     * @param lastAmbientTemperature
     *   Ambient temperature valid up until (not including) the current tick
-    * @param state
-    *   Current state of the houses
+    * @param lastThermalGridState
+    *   state of the thermalGrid until this tick
+    * @param isRunning
+    *   determines whether the heat pump is running or not
     * @param qDot
-    *   Infeed to the grid
+    *   Infeed to the grid from thermal generation (e.g. heat pump) or thermal
+    *   storages
+    * @param thermalDemands
+    *   holds the thermal demands of the thermal units (house, storage)
     * @return
-    *   Updated thermal grid state
+    *   Updated thermal grid state and the thermalThreshold if there is one
     */
   private def handleInfeed(
       relevantData: HpRelevantData,
       lastAmbientTemperature: Temperature,
-      state: ThermalGridState,
+      lastThermalGridState: ThermalGridState,
+      isRunning: Boolean,
       qDot: Power,
-  ): (ThermalGridState, Option[ThermalThreshold]) =
-    house.zip(state.houseState) match {
-      case Some((thermalHouse, lastHouseState)) =>
-        /* Set thermal power exchange with storage to zero */
-        // TODO: We would need to issue a storage result model here...
-        val updatedStorageState = storage.zip(state.storageState) match {
-          case Some((thermalStorage, storageState)) =>
-            Some(
-              thermalStorage
-                .updateState(
-                  relevantData.currentTick,
-                  zeroKW,
-                  storageState,
-                )
-                ._1
-            )
-          case _ => state.storageState
-        }
+      thermalDemands: ThermalDemandWrapper,
+  ): (ThermalGridState, Option[ThermalThreshold]) = {
+    // TODO: We would need to issue a storage result model here...
 
-        val (updatedHouseState, maybeHouseThreshold) =
-          thermalHouse.determineState(
-            relevantData,
-            lastHouseState,
-            lastAmbientTemperature,
-            qDot,
+    /* Consider the action in the last state */
+    val qDotHouseLastState =
+      lastThermalGridState.houseState.map(_.qDot).getOrElse(zeroKW)
+    val qDotStorageLastState =
+      lastThermalGridState.storageState.map(_.qDot).getOrElse(zeroKW)
+
+    // We can use the qDots from lastState to keep continuity. If...
+    if (
+      // ... house was heated in lastState but not from Storage and has still some demand.
+      ((qDotHouseLastState > zeroKW && (qDotStorageLastState >= zeroKW) && thermalDemands.houseDemand.hasAdditionalDemand) ||
+      // ... storage was filled up in the lastState and has still additional demand
+      // But only if the house not reached some requiredDemand.
+      qDotStorageLastState > zeroKW && thermalDemands.heatStorageDemand.hasAdditionalDemand && !thermalDemands.houseDemand.hasRequiredDemand)
+    ) {
+      // We can continue for the house
+      val (updatedHouseState, thermalHouseThreshold, remainingQDotHouse) =
+        handleInfeedHouse(
+          relevantData,
+          lastAmbientTemperature,
+          lastThermalGridState,
+          qDotHouseLastState,
+        )
+
+      // ...and for the storage
+      val (updatedStorageState, thermalStorageThreshold) = {
+        // In case the ThermalHouse could not handle the infeed it will be used for the storage.
+        if (remainingQDotHouse > qDotStorageLastState) {
+          handleStorageCases(
+            relevantData.currentTick,
+            lastThermalGridState,
+            remainingQDotHouse,
           )
+        } else {
+          handleStorageCases(
+            relevantData.currentTick,
+            lastThermalGridState,
+            qDotStorageLastState,
+          )
+        }
+      }
 
+      val nextThreshold = determineMostRecentThreshold(
+        thermalHouseThreshold,
+        thermalStorageThreshold,
+      )
+      (
+        lastThermalGridState.copy(
+          houseState = updatedHouseState,
+          storageState = updatedStorageState,
+        ),
+        nextThreshold,
+      )
+    }
+    // Handle edge case where house was heated from storage...
+    else if (qDotHouseLastState > zeroKW && qDotStorageLastState < zeroKW) {
+      // ...and HP gets activated in current tick
+      if (isRunning) {
+        handleCases(
+          relevantData,
+          lastAmbientTemperature,
+          lastThermalGridState,
+          qDot,
+          zeroKW,
+        )
+      } else {
+        // ... or continue lastState's behaviour
+        handleCases(
+          relevantData,
+          lastAmbientTemperature,
+          lastThermalGridState,
+          qDotHouseLastState,
+          qDotStorageLastState,
+        )
+      }
+    }
+    // Handle edge case where house should be heated from storage
+    else if (!isRunning && qDot > zeroKW) {
+      handleCases(
+        relevantData,
+        lastAmbientTemperature,
+        lastThermalGridState,
+        qDot,
+        -qDot,
+      )
+    }
+    // or finally check for all other cases.
+    else
+      handleFinalInfeedCases(
+        thermalDemands,
+        relevantData,
+        lastAmbientTemperature,
+        lastThermalGridState,
+        qDot,
+      )
+  }
+
+  /** Handles the last cases of [[ThermalGrid.handleInfeed]], where the thermal
+    * infeed should be determined.
+    *
+    * | house req. demand | house add. demand | storage req. demand | storage add. demand | qDot to house | qDot to storage |
+    * |:------------------|:------------------|:--------------------|:--------------------|:--------------|:----------------|
+    * | true              | true              | true                | true                | true          | false           |
+    * | true              | true              | true                | false               | true          | false           |
+    * | true              | true              | false               | true                | true          | false           |
+    * | true              | true              | false               | false               | true          | false           |
+    * | true              | false             | true                | true                | true          | false           |
+    * | true              | false             | true                | false               | true          | false           |
+    * | true              | false             | false               | true                | true          | false           |
+    * | true              | false             | false               | false               | true          | false           |
+    * | false             | true              | true                | true                | false         | true            |
+    * | false             | true              | true                | false               | false         | true            |
+    * | false             | true              | false               | true                | false         | true            |
+    * | false             | true              | false               | false               | true          | false           |
+    * | false             | false             | true                | true                | false         | true            |
+    * | false             | false             | true                | false               | false         | true            |
+    * | false             | false             | false               | true                | false         | true            |
+    * | false             | false             | false               | false               | false         | false           |
+    *
+    * This can be simplified to four cases
+    * | No | Conditions                           | Result    |
+    * |:---|:-------------------------------------|:----------|
+    * | 1  | if house.reqD                        | house     |
+    * | 2  | else if storage.reqD OR storage.addD | storage   |
+    * | 3  | else if house.addD                   | house     |
+    * | 4  | else                                 | no output |
+    *
+    * @param thermalDemands
+    *   holds the thermal demands of the thermal units (house, storage)
+    * @param relevantData
+    *   data of heat pump including state of the heat pump
+    * @param lastAmbientTemperature
+    *   Ambient temperature valid up until (not including) the current tick
+    * @param gridState
+    *   Current state of the thermalGrid
+    * @param qDot
+    *   Infeed to the grid from thermal generation (e.g. heat pump) or thermal
+    *   storages
+    * @return
+    *   Updated thermal grid state and the thermalThreshold if there is one
+    */
+  private def handleFinalInfeedCases(
+      thermalDemands: ThermalDemandWrapper,
+      relevantData: HpRelevantData,
+      lastAmbientTemperature: Temperature,
+      gridState: ThermalGridState,
+      qDot: Power,
+  ): (ThermalGridState, Option[ThermalThreshold]) = {
+
+    if (thermalDemands.houseDemand.hasRequiredDemand)
+      handleCases(
+        relevantData,
+        lastAmbientTemperature,
+        gridState,
+        qDot,
+        zeroKW,
+      )
+    else if (
+      thermalDemands.heatStorageDemand.hasRequiredDemand || thermalDemands.heatStorageDemand.hasAdditionalDemand
+    )
+      handleCases(
+        relevantData,
+        lastAmbientTemperature,
+        gridState,
+        zeroKW,
+        qDot,
+      )
+    else if (thermalDemands.houseDemand.hasAdditionalDemand)
+      handleCases(
+        relevantData,
+        lastAmbientTemperature,
+        gridState,
+        qDot,
+        zeroKW,
+      )
+    else
+      handleCases(
+        relevantData,
+        lastAmbientTemperature,
+        gridState,
+        zeroKW,
+        zeroKW,
+      )
+
+  }
+
+  /** Handles the different cases, of thermal flows from and into the thermal
+    * grid.
+    *
+    * @param relevantData
+    *   data of heat pump including state of the heat pump
+    * @param lastAmbientTemperature
+    *   Ambient temperature until this tick
+    * @param state
+    *   Current state of the thermal grid
+    * @param qDotHouse
+    *   Infeed to the house
+    * @param qDotHeatStorage
+    *   Infeed to the heat storage (positive: Storage is charging, negative:
+    *   Storage is discharging)
+    * @return
+    *   Updated thermal grid state and the next threshold if there is one
+    */
+  private def handleCases(
+      relevantData: HpRelevantData,
+      lastAmbientTemperature: Temperature,
+      state: ThermalGridState,
+      qDotHouse: Power,
+      qDotHeatStorage: Power,
+  ): (ThermalGridState, Option[ThermalThreshold]) = {
+    val (updatedHouseState, thermalHouseThreshold, _) =
+      handleInfeedHouse(
+        relevantData,
+        lastAmbientTemperature,
+        state,
+        qDotHouse,
+      )
+
+    val (updatedStorageState, thermalStorageThreshold) =
+      handleStorageCases(relevantData.currentTick, state, qDotHeatStorage)
+
+    val nextThreshold = determineMostRecentThreshold(
+      thermalHouseThreshold,
+      thermalStorageThreshold,
+    )
+
+    (
+      state.copy(
+        houseState = updatedHouseState,
+        storageState = updatedStorageState,
+      ),
+      nextThreshold,
+    )
+  }
+
+  /** Handles the case, when the house has heat demand and will be heated up
+    * here.
+    *
+    * @param relevantData
+    *   data of heat pump including state of the heat pump
+    * @param lastAmbientTemperature
+    *   Ambient temperature until this tick
+    * @param state
+    *   Current state of the houses
+    * @param qDotHouse
+    *   Infeed into the house
+    * @return
+    *   Updated thermal house state, a ThermalThreshold and the remaining qDot
+    */
+  private def handleInfeedHouse(
+      relevantData: HpRelevantData,
+      lastAmbientTemperature: Temperature,
+      state: ThermalGridState,
+      qDotHouse: Power,
+  ): (Option[ThermalHouseState], Option[ThermalThreshold], Power) = {
+    (house, state.houseState) match {
+      case (Some(thermalHouse), Some(lastHouseState)) =>
+        val (newState, threshold) = thermalHouse.determineState(
+          relevantData,
+          lastHouseState,
+          lastAmbientTemperature,
+          qDotHouse,
+        )
+        /* Check if house can handle the thermal feed in */
         if (
           thermalHouse.isInnerTemperatureTooHigh(
-            updatedHouseState.innerTemperature
+            newState.innerTemperature
           )
         ) {
-          /* The house is already heated up fully, set back the infeed and put it into storage, if available */
           val (fullHouseState, maybeFullHouseThreshold) =
             thermalHouse.determineState(
               relevantData,
@@ -230,63 +495,53 @@ final case class ThermalGrid(
               lastAmbientTemperature,
               zeroKW,
             )
-          storage.zip(updatedStorageState) match {
-            case Some((thermalStorage, storageState)) =>
-              val (updatedStorageState, maybeStorageThreshold) =
-                thermalStorage.updateState(
-                  relevantData.currentTick,
-                  qDot,
-                  storageState,
-                )
-
-              /* Both house and storage are updated. Determine what reaches the next threshold */
-              val nextThreshold = determineMostRecentThreshold(
-                maybeFullHouseThreshold,
-                maybeStorageThreshold,
-              )
-
-              (
-                state.copy(
-                  houseState = Some(fullHouseState),
-                  storageState = Some(updatedStorageState),
-                ),
-                nextThreshold,
-              )
-            case None =>
-              /* There is no storage, house determines the next activation */
-              (
-                state.copy(houseState = Some(fullHouseState)),
-                maybeFullHouseThreshold,
-              )
-          }
+          (Some(fullHouseState), maybeFullHouseThreshold, qDotHouse)
         } else {
-          /* The house can handle the infeed */
-          (
-            state.copy(houseState = Some(updatedHouseState)),
-            maybeHouseThreshold,
-          )
+          (Some(newState), threshold, zeroKW)
         }
-
-      case None =>
-        storage.zip(state.storageState) match {
-          case Some((thermalStorage, storageState)) =>
-            val (updatedStorageState, maybeStorageThreshold) =
-              thermalStorage.updateState(
-                relevantData.currentTick,
-                qDot,
-                storageState,
-              )
-            (
-              state.copy(storageState = Some(updatedStorageState)),
-              maybeStorageThreshold,
-            )
-          case None =>
-            throw new InconsistentStateException(
-              "A thermal grid has to contain either at least a house or a storage."
-            )
-        }
+      case _ => (None, None, zeroKW)
     }
+  }
 
+  /** Handles the cases, when the storage has heat demand and will be filled up
+    * here (positive qDot) or will return its stored energy into the thermal
+    * grid (negative qDot).
+    * @param tick
+    *   Current tick
+    * @param state
+    *   Current state of the houses
+    * @param qDotStorage
+    *   Infeed to the storage (positive: Storage is charging, negative: Storage
+    *   is discharging)
+    * @return
+    *   Updated thermal grid state
+    */
+  private def handleStorageCases(
+      tick: Long,
+      state: ThermalGridState,
+      qDotStorage: Power,
+  ): (Option[ThermalStorageState], Option[ThermalThreshold]) = {
+    (storage, state.storageState) match {
+      case (Some(thermalStorage), Some(lastStorageState)) =>
+        val (newState, threshold) = thermalStorage.updateState(
+          tick,
+          qDotStorage,
+          lastStorageState,
+        )
+        (Some(newState), threshold)
+      case _ => (None, None)
+    }
+  }
+
+  /** Determines the most recent threshold of two given input thresholds
+    *
+    * @param maybeHouseThreshold
+    *   Option of a possible next threshold of the thermal house
+    * @param maybeStorageThreshold
+    *   Option of a possible next threshold of the thermal storage
+    * @return
+    *   The next threshold
+    */
   private def determineMostRecentThreshold(
       maybeHouseThreshold: Option[ThermalThreshold],
       maybeStorageThreshold: Option[ThermalThreshold],
@@ -308,34 +563,37 @@ final case class ThermalGrid(
     *   data of heat pump including state of the heat pump
     * @param lastAmbientTemperature
     *   Ambient temperature valid up until (not including) the current tick
-    * @param state
-    *   Current state of the houses
+    * @param lastThermalGridState
+    *   state of the thermalGrid until this tick
     * @param qDot
-    *   Infeed to the grid
+    *   Infeed to the grid from thermal generation (e.g. heat pump) or thermal
+    *   storages
     * @return
     *   Updated thermal grid state
     */
   private def handleConsumption(
       relevantData: HpRelevantData,
       lastAmbientTemperature: Temperature,
-      state: ThermalGridState,
+      lastThermalGridState: ThermalGridState,
       qDot: Power,
   ): (ThermalGridState, Option[ThermalThreshold]) = {
     /* House will be left with no influx in all cases. Determine if and when a threshold is reached */
     val maybeUpdatedHouseState =
-      house.zip(state.houseState).map { case (house, houseState) =>
-        house.determineState(
-          relevantData,
-          houseState,
-          lastAmbientTemperature,
-          zeroMW,
-        )
+      house.zip(lastThermalGridState.houseState).map {
+        case (house, houseState) =>
+          house.determineState(
+            relevantData,
+            houseState,
+            lastAmbientTemperature,
+            zeroMW,
+          )
       }
 
     /* Update the state of the storage */
     val maybeUpdatedStorageState =
-      storage.zip(state.storageState).map { case (storage, storageState) =>
-        storage.updateState(relevantData.currentTick, qDot, storageState)
+      storage.zip(lastThermalGridState.storageState).map {
+        case (storage, storageState) =>
+          storage.updateState(relevantData.currentTick, qDot, storageState)
       }
 
     val (revisedHouseState, revisedStorageState) =
@@ -343,8 +601,8 @@ final case class ThermalGrid(
         relevantData,
         maybeUpdatedHouseState,
         maybeUpdatedStorageState,
-        state.houseState,
-        state.storageState,
+        lastThermalGridState.houseState,
+        lastThermalGridState.storageState,
         lastAmbientTemperature,
         qDot,
       )
@@ -355,7 +613,7 @@ final case class ThermalGrid(
     )
 
     (
-      state.copy(
+      lastThermalGridState.copy(
         houseState = revisedHouseState.map(_._1),
         storageState = revisedStorageState.map(_._1),
       ),
@@ -381,7 +639,8 @@ final case class ThermalGrid(
     * @param lastAmbientTemperature
     *   Ambient temperature valid up until (not including) the current tick
     * @param qDot
-    *   Thermal influx
+    *   Infeed to the grid from thermal generation (e.g. heat pump) or thermal
+    *   storages
     * @return
     *   Options to revised thermal house and storage state
     */
@@ -561,7 +820,11 @@ object ThermalGrid {
     * than or equal to the absolutely required energy. Thus, this class can only
     * be instantiated via factory.
     * @param required
-    *   The absolutely required energy to reach target state
+    *   The absolutely required energy to reach target state. For
+    *   [[ThermalHouse]] this would be the energy demand to reach the boundary
+    *   or targetTemperature. For [[ThermalStorage]] this would be the amount of
+    *   energy to get fully charged when empty. If the [[ThermalStorage]] is not
+    *   empty, the required energy is zero.
     * @param possible
     *   The maximum possible energy, that can be handled
     */
@@ -581,8 +844,7 @@ object ThermalGrid {
   object ThermalEnergyDemand {
 
     /** Builds a new instance of [[ThermalEnergyDemand]]. If the possible energy
-      * is less than the required energy, this is considered to be a bad state
-      * and the required energy is curtailed to the possible energy.
+      * is less than the required energy, this is considered to be a bad state.
       * @param required
       *   The absolutely required energy to reach target state
       * @param possible
@@ -594,10 +856,19 @@ object ThermalGrid {
         required: Energy,
         possible: Energy,
     ): ThermalEnergyDemand = {
-      if (possible < required)
-        new ThermalEnergyDemand(possible, possible)
-      else
-        new ThermalEnergyDemand(required, possible)
+      if (
+        math.abs(possible.toKilowattHours) < math.abs(required.toKilowattHours)
+      )
+        throw new InvalidParameterException(
+          s"The possible amount of energy $possible is smaller than the required amount of energy $required. This is not supported."
+        )
+
+      if (possible.toKilowattHours < 0 || required.toKilowattHours < 0)
+        throw new InvalidParameterException(
+          s"The possible $possible or required $required amount of energy cannot be negative. This is not supported."
+        )
+
+      new ThermalEnergyDemand(required, possible)
     }
 
     def noDemand: ThermalEnergyDemand = ThermalEnergyDemand(